refactor: move clearPlot to its own file
