+++ src/game-logic/reducers/clearPlot.js
