+++ src/game-logic/reducers/index.test.js
@@ -34,7 +34,7 @@ import {
 import { huggingMachine, sampleCropItem1 } from '../../data/items'
 import { sampleRecipe1 } from '../../data/recipes'
 import { itemsMap } from '../../data/maps'
-import { genders, standardCowColors, toolType, toolLevel } from '../../enums'
+import { genders, standardCowColors, toolLevel } from '../../enums'
 import {
   farmProductSalesVolumeNeededForLevel,
   generateCow,
@@ -44,7 +44,6 @@ import {
   getCowValue,
   getPlotContentFromItemId,
   getPriceEventForCrop,
-  isRandomNumberLessThan,
 } from '../../utils'
 
 import * as fn from './'
@@ -1943,117 +1942,6 @@ describe('changeCowBreedingPenResident', () => {
             daysUntilBirth: COW_GESTATION_PERIOD_DAYS,
           },
         })
-      })
-    })
-  })
-})
-
-describe('clearPlot', () => {
-  describe('plotContent is a crop', () => {
-    test('clears the plot', () => {
-      const { field } = fn.clearPlot(
-        {
-          field: [[testCrop({ itemId: 'sample-crop-1' })]],
-          toolLevels: { [toolType.HOE]: toolLevel.DEFAULT },
-          inventory: [],
-          inventoryLimit: -1,
-        },
-        0,
-        0
-      )
-
-      expect(field[0][0]).toBe(null)
-    })
-
-    describe('there is no room in the inventory', () => {
-      test('clears the plot', () => {
-        const { field, inventory } = fn.clearPlot(
-          {
-            field: [[testCrop({ itemId: 'sample-crop-1' })]],
-            toolLevels: { [toolType.HOE]: toolLevel.DEFAULT },
-            inventory: [{ id: 'sample-item-1', quantity: 5 }],
-            inventoryLimit: 5,
-          },
-          0,
-          0
-        )
-
-        expect(field[0][0]).toBe(null)
-        expect(inventory).toEqual([{ id: 'sample-item-1', quantity: 5 }])
-      })
-    })
-  })
-
-  describe('crop is fully grown', () => {
-    test('harvests crop', () => {
-      const { field, inventory } = fn.clearPlot(
-        {
-          field: [[testCrop({ itemId: 'sample-crop-1', daysWatered: 3 })]],
-          toolLevels: { [toolType.HOE]: toolLevel.DEFAULT },
-          inventory: [],
-          inventoryLimit: 10,
-        },
-        0,
-        0
-      )
-
-      expect(field[0][0]).toBe(null)
-      expect(inventory).toEqual([{ id: 'sample-crop-1', quantity: 1 }])
-    })
-  })
-
-  describe('plotContent is replantable', () => {
-    test('updates state', () => {
-      const { field, inventory } = fn.clearPlot(
-        {
-          field: [[getPlotContentFromItemId('replantable-item')]],
-          toolLevels: { [toolType.HOE]: toolLevel.DEFAULT },
-          inventory: [],
-          inventoryLimit: -1,
-        },
-        0,
-        0
-      )
-
-      expect(inventory).toEqual([{ id: 'replantable-item', quantity: 1 }])
-      expect(field[0][0]).toBe(null)
-    })
-
-    describe('there is no room in the inventory', () => {
-      test('no-ops', () => {
-        const inputState = {
-          field: [[getPlotContentFromItemId('replantable-item')]],
-          toolLevels: { [toolType.HOE]: toolLevel.DEFAULT },
-          inventory: [{ id: 'sample-item-1', quantity: 5 }],
-          inventoryLimit: 5,
-        }
-        const state = fn.clearPlot(inputState, 0, 0)
-
-        expect(state).toEqual(inputState)
-      })
-    })
-  })
-
-  describe('hoe upgrades', () => {
-    beforeEach(() => {
-      isRandomNumberLessThan.mockReturnValue(true)
-    })
-
-    describe('crop is not fully grown', () => {
-      test('returns seed to inventory', () => {
-        const { field, inventory } = fn.clearPlot(
-          {
-            field: [[testCrop({ itemId: 'sample-crop-1' })]],
-            toolLevels: { [toolType.HOE]: toolLevel.BRONZE },
-            inventory: [],
-            inventoryLimit: 10,
-          },
-          0,
-          0
-        )
-
-        expect(field[0][0]).toBe(null)
-        expect(inventory).toEqual([{ id: 'sample-crop-seeds-1', quantity: 1 }])
       })
     })
   })
